+++ tests/fast-integration/kyma.js
@@ -1,9 +1,5 @@
 #!/usr/bin/env node
 const installer = require("./installer")
-const axios = require("axios");
-var AdmZip = require('adm-zip');
-var fs = require('fs');
-const { join } = require("path");
 const { switchDebug } = require("./utils");
 
 function installOptions(yargs) {
@@ -14,8 +10,8 @@ function installOptions(yargs) {
           - To use a specific release, write "kyma install --source=1.15.1".\n\
           - To use the local sources, write "kyma install --source=local".'
     },
-    'skip-modules': {
-      describe: 'Skip modules from the list (comma separated)'
+    'skip-components': {
+      describe: 'Skip components from the list (comma separated)'
     },
     'new-eventing' : {
       describe: 'Install new eventing instead of knative'
@@ -45,7 +41,6 @@ const argv = require('yargs/yargs')(process.argv.slice(2))
   .options({ 'verbose': { alias: 'v', describe: 'Displays details of actions triggered by the command.' } })
   .command('install', 'Installs Kyma on a running Kubernetes cluster', installOptions, install)
   .command('uninstall', 'Removes Kyma from cluster', uninstallOptions, uninstall)
-  .command('download', 'Downloads Kyma sources into ./tmp folder', installOptions, download)
   .demandCommand(1, 1, 'Command is missing')
   .example('$0 install --skip-modules=monitoring,tracing,kiali')
   .strict()
@@ -57,10 +52,11 @@ async function install(argv) {
   let src = undefined;
   verbose(argv);
   if (argv.source) {
-    await download(argv)
-    src = join(__dirname, './tmp/resources')
+    src = await installer.downloadCharts(argv)
   }
-  await installer.installKyma(src, "1.8.2", "", false, argv);
+  const skipComponents = argv.skipComponents ? argv.skipComponents.split(',').map(c => c.trim()) : [];
+
+  await installer.installKyma({resourcesPath:src,skipComponents});
   console.log('Kyma installed')
 }
 
@@ -70,41 +66,3 @@ async function uninstall(argv) {
   console.log('Kyma uninstalled')
 }
 
-
-async function downloadFile(url, filename) {
-  const writer = fs.createWriteStream(filename)
-
-  const response = await axios.get(url, { responseType: 'stream' });
-
-  response.data.pipe(writer)
-
-  return new Promise((resolve, reject) => {
-    writer.on('finish', resolve)
-    writer.on('error', reject)
-  })
-}
-
-
-async function download(argv) {
-  verbose(argv);
-  var dir = join(__dirname, './tmp');
-
-  if (fs.existsSync(dir)) {
-    fs.rmdirSync(dir, { recursive: true })
-  }
-  fs.mkdirSync(dir);
-
-  const repo = 'kyma-project/kyma'
-  const branch = argv.source || 'master';
-  const zipFile = join(__dirname, './tmp/', branch + '.zip')
-  await downloadFile(`https://codeload.github.com/${repo}/zip/${branch}`, zipFile)
-  var zip = new AdmZip(zipFile);
-  var zipEntries = zip.getEntries();
-  zipEntries.forEach(function (zipEntry) {
-    const target = zipEntry.entryName.split('/').slice(1).join('/');
-    if (target == 'resources/') {
-      zip.extractEntryTo(zipEntry.entryName, dir);
-      fs.renameSync(join(dir, zipEntry.entryName), join(dir, 'resources'));
-    }
-  });
-}
