+++ tests/fast-integration/test/1-install.js
@@ -4,6 +4,6 @@ describe("Installation", function () {
   this.timeout(10 * 60 * 1000);
 
   it("Kyma should successfully install", async function () {
-    await installKyma(undefined, "1.8.2", "");
+    await installKyma();
   });
 });
